move powgen workflow to seaprate file
