+++ src/ess/snspowder/powgen/__init__.py
@@ -7,10 +7,6 @@ Note that this module is temporary and will be removed in favor of
 the ``dream`` module when that is available.
 """
 
-import sciline
-
-from ess.powder import providers as powder_providers
-from ess.powder.types import NeXusDetectorName
 from . import beamline, data
 from .calibration import load_calibration
 from .instrument_view import instrument_view
@@ -22,24 +18,4 @@ providers = (
 """Sciline Providers for POWGEN-specific functionality."""
 
 
-def default_parameters() -> dict:
-    return {NeXusDetectorName: "powgen_detector"}
-
-
-def PowgenWorkflow() -> sciline.Pipeline:
-    """
-    Workflow with default parameters for the Powgen SNS instrument.
-    """
-    return sciline.Pipeline(
-        providers=powder_providers + providers, params=default_parameters()
-    )
-
-
-__all__ = [
-    'PowgenWorkflow',
-    'beamline',
-    'data',
-    'default_parameters',
-    'instrument_view',
-    'load_calibration',
-]
+__all__ = ['beamline', 'data', 'instrument_view', 'load_calibration', 'providers']
